remote_new_download/requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrdict==2.0.1
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
babelfish==0.5.5
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.4
commonmark==0.9.1
comtypes==1.1.7
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
Cython==0.29.22
Deprecated==1.2.11
distlib==0.3.1
Django==3.1.3
dominate==2.6.0
EasyProcess==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
get-video-properties==0.1.1
guessit==3.1.1
hyperlink==21.0.0
idm==0.39
idna==2.8
image==1.5.33
IMDbPY==2020.9.25
incremental==21.3.0
install==1.3.4
internetdownloadmanager==0.0.2
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector==2.2.9
ndg-httpsclient==0.5.1
numpy==1.20.1
olefile==0.46
openpyxl==3.0.5
parsel==1.6.0
pefile==2019.4.18
Pillow==8.1.2
pipenv==2020.11.15
pornhubapi==0.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyGetWindow==0.0.9
PyGithub==1.54.1
Pygments==2.7.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==1.7.1
pylint==2.6.0
PyMsgBox==1.0.9
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
pyTelegramBotAPI==3.7.5
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.11.0
python-qbittorrent==0.4.2
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.4
PyVirtualDisplay==1.3.2
pywin32==300
pywin32-ctypes==0.2.0
qbittorrent-api==2020.12.15
queuelib==1.5.0
RarbgAPI==0.4.2
rebulk==2.0.1
requests==2.25.0
rich==9.8.2
Scrapy==1.4.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
sslcommerz-lib==1.0
tesseract==0.1.3
toml==0.10.2
tqdm==4.53.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.2.1
virtualenv-clone==0.5.4
visitor==0.1.3
w3lib==1.22.0
Werkzeug==1.0.0
win10toast==0.9
windows-curses==2.2.0
wrapt==1.12.1
WTForms==2.2.1
zope.interface==5.2.0
